requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
psycopg2-binary==2.9.7
openpyxl==3.1.2
scikit-learn==1.3.2
numpy==1.24.4